Rename bin/ to filecheck/
